requirements.txt
fastmcp==0.2.0
openai==1.51.2
pydantic==2.9.2
pillow==10.4.0
python-dotenv==1.0.1
uvicorn==0.31.1
aiofiles==24.1.0
